# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IAmSurfer API",
    "version": "1",
    "description": "API REST versionada (/api/v1) do app mobile IAmSurfer. Autenticação via JWT Bearer (access 15min + refresh 30d rotativo). Erros no formato { \"error\": { \"code\", \"message\", \"fields\" } }. Feeds paginados por cursor: { \"items\": [...], \"next_cursor\": ... }."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/stories": {
      "get": {
        "operationId": "get_api_stories_api_story_bar",
        "tags": [
          "stories"
        ],
        "summary": "Story bar agrupada por usuário (mesmo comportamento do site).",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "operationId": "post_api_stories_api_create_story",
        "tags": [
          "stories"
        ],
        "summary": "Cria um story a partir de uma mídia já enviada ao Cloudinary (A11).",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/stories/{story_id}/seen": {
      "post": {
        "operationId": "post_api_stories_api_mark_seen",
        "tags": [
          "stories"
        ],
        "summary": "Marca um story como visto (idempotente).",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/stories/{story_id}": {
      "delete": {
        "operationId": "delete_api_stories_api_delete_story",
        "tags": [
          "stories"
        ],
        "summary": "Apaga um story próprio (também remove a mídia do Cloudinary).",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Formato de erro único da API.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Código estável, legível por máquina."
              },
              "message": {
                "type": "string",
                "description": "Mensagem humana (pt-BR)."
              },
              "fields": {
                "type": "object",
                "description": "Erros de validação por campo.",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  }
}